requirements.txt
flask
flask-cors
google-genai
networkx
dotenv
logging
pathlib